+++ src/data/rito.js
@@ -1,15 +1,21 @@
 const rito = {
     id: 'rito',
+    behanceLink: 'https://behance.net',
     en: {
         title: 'Rito',
         subtitle: 'web/app editorial project',
-        overview:
-            'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers.',
-        focus: 'Experience and Interface Design',
+        overview: [
+            'Project created from scratch, in 4 days, in collaboration with Tadeau Carabias.',
+            'The challenge was to create a responsive online platform for a magazine to meet the needs and goals of users. The goal was to deliver a consistent, branded, and engaging experience and our main focus were on UI Design.'
+        ],
+        scope: 'User Research, analyzing, user flow, information architecture, wireframing, mock-up, design system, prototyping, and usability testing.',
         categories: 'Website/Mobile ',
-        brief: 'Rito is a collaborative project,  done with Tadeu Carabias in  is a  magazine oriented to a young audience. Our public intended to be an audience who enjoys knowing new and emergent artists, who like to know what happens in the artistic environment, is aware of new trends, and receives recommendations on new artistic creations to continue feeding their artistic interest spirit.',
-        research:
-            'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
+        brief: 'Rito is an emergent art magazine to promote new and upcoming artists and review the work of established ones, with the intent to democratize art and enable easier access for everyone both artist and viewers',
+        research: [
+            'Starting our creative process by choosing an already created user persona (Paula) allowed us to think about this editorial project, which mutated and evolved from the beginning excitingly and organically. By having specific characteristics about Paula, we could orient our research.',
+            'At first, we searched the web for different digital and physical magazines to understand how this type of market works and what people want to consume when they decide to read an art magazine.',
+            'After visiting and comparing different sites, the information they offer, how content is organized, their aesthetics, their accessibility, and even the frequency of content, we started thinking of questions for the interviews that would allow us to achieve our goal.'
+        ],
         visualConcept:
             'Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga. Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.Lorem ipsum dolor sit amet consectetur adipisicing elit. Id praesentium atque ipsam debitis dolorem ullam maxime tenetur esse consectetur perspiciatis amet, nam aspernatur quis adipisci culpa modi laborum incidunt fuga.',
         prototype: